pyproject.toml
[tool.poetry]
name = "Diploma Blockchain"
version = "0.1.0"
description = "une blockchain post quantique en Python"
authors = ["Julien Denizot <luunz71@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
pypqc = "0.0.6.2"
flake8 = "^7.1.1"
pycryptodome = "^3.21.0"
rich = "^13.9.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
